+++ packages/chains/src/filecoin.ts
@@ -1,0 +1,21 @@
+import { Chain } from './types'
+
+export const filecoin: Chain = {
+  id: 314,
+  name: 'Filecoin Mainnet',
+  network: 'filecoin-mainnet',
+  nativeCurrency: {
+    decimals: 18,
+    name: 'filecoin',
+    symbol: 'FIL',
+  },
+  rpcUrls: {
+    default: { http: ['https://api.node.glif.io'] },
+    public: { http: ['https://api.node.glif.io'] },
+  },
+  blockExplorers: {
+    default: { name: 'Filfox', url: 'https://filfox.info/en' },
+    filscan: { name: 'Filscan', url: 'https://filscan.io' },
+    filscout: { name: 'Filscout', url: 'https://filscout.io/en' },
+  },
+}