+++ packages/chains/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wagmi/chains
 
+## 0.1.13
+
+### Patch Changes
+
+- Added Filecoin & Filecoin Hyperspace chains.
+
+## 0.1.12
+
+### Patch Changes
+
+- 72a9539: Added Filecoin & Filecoin Hyperspace chains.
+
 ## 0.1.11
 
 ### Patch Changes